# app/Http/Resources/Admin/Admins/AdminResource.php
<?php

namespace App\Http\Resources\Admin\Admins;

use Illuminate\Http\Request;
use Illuminate\Http\Resources\Json\JsonResource;
use OpenApi\Annotations as OA;

/**
 * @property mixed $username
 * @property mixed $id
 * @property mixed $email
 * @property mixed $name
 * @property mixed $surname
 * @property mixed $roles
 *
 * @OA\Schema(
 *     schema="AdminResource",
 *     type="object",
 *     title="Admin Resource",
 *     description="Admin resource representation",
 *     @OA\Property(property="id", type="integer", example=1),
 *     @OA\Property(property="username", type="string", example="admin"),
 *     @OA\Property(property="email", type="string", example="admin@example.com"),
 *     @OA\Property(property="name", type="string", example="John"),
 *     @OA\Property(property="surname", type="string", example="Doe"),
 *     @OA\Property(property="roles", type="array", @OA\Items(type="integer", example=1))
 * )
 */
class AdminResource extends JsonResource
{
    /**
     * Transform the resource into an array.
     *
     * @param Request $request
     * @return array
     */

    public function toArray(Request $request): array
    {
        return [
            'id' => $this->id,
            'username' => $this->username,
            'email' => $this->email,
            'name' => $this->name,
            'surname' => $this->surname,
            'roles' => data_get($this->roles, '*.id'),
        ];
    }
}