# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FAQ API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/control/admins/load": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get list of admins",
                "operationId": "7e4cd459a2d619299c9e90c017fa9dd0",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AdminsResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiToken": [],
                        "SanctumBearerToken": []
                    }
                ]
            }
        },
        "/api/control/admins/show/{id}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get admin by ID",
                "operationId": "ff8354912b63b88bfbf2648a59dd7057",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdminResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiToken": [],
                        "SanctumBearerToken": []
                    }
                ]
            }
        },
        "/api/control/admins/add": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Create a new admin",
                "operationId": "bebd869a85fa7cbfaac9e6642ffde9e2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AdminStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Admin created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdminsResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiToken": [],
                        "SanctumBearerToken": []
                    }
                ]
            }
        },
        "/api/control/admins/update/{id}": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Update an existing admin",
                "operationId": "6d055b16d60275095c549f92a54095da",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AdminUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Admin updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdminsResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiToken": [],
                        "SanctumBearerToken": []
                    }
                ]
            }
        },
        "/api/control/admins/delete/{id}": {
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Delete an admin",
                "operationId": "3802e98297faf93823280932475ae113",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Admin deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiToken": [],
                        "SanctumBearerToken": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AdminStoreRequest": {
                "title": "Admin Store Request",
                "description": "Request body for creating a new admin",
                "required": [
                    "username",
                    "email",
                    "password",
                    "roles"
                ],
                "properties": {
                    "username": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "admin"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 150,
                        "example": "admin@example.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "P@ssw0rd!"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "John"
                    },
                    "surname": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Doe"
                    }
                },
                "type": "object"
            },
            "AdminUpdateRequest": {
                "title": "Admin Update Request",
                "description": "Request body for updating an existing admin",
                "required": [
                    "username",
                    "email",
                    "roles"
                ],
                "properties": {
                    "username": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "admin"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 150,
                        "example": "admin@example.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "P@ssw0rd!"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "John"
                    },
                    "surname": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Doe"
                    }
                },
                "type": "object"
            },
...
            "AdminResource": {
                "title": "Admin Resource",
                "description": "Admin resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "username": {
                        "type": "string",
                        "example": "admin"
                    },
                    "email": {
                        "type": "string",
                        "example": "admin@example.com"
                    },
                    "name": {
                        "type": "string",
                        "example": "John"
                    },
                    "surname": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                },
                "type": "object"
            },
            "AdminsResource": {
                "title": "Admins Resource",
                "description": "Admins resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "username": {
                        "type": "string",
                        "example": "admin"
                    },
                    "email": {
                        "type": "string",
                        "example": "admin@example.com"
                    },
                    "name": {
                        "type": "string",
                        "example": "John"
                    },
                    "surname": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RolesListResource"
                        }
                    },
                    "created_user": {
                        "type": "string",
                        "example": "creator"
                    },
                    "created_date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-01T12:00:00Z"
                    }
                },
                "type": "object"
            },
            "RolesListResource": {
                "title": "Roles List Resource",
                "description": "Roles list resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Admin"
                    },
                    "guard_name": {
                        "type": "string",
                        "example": "web"
                    }
                },
                "type": "object"
            },
...
            "GeneralResource": {
                "title": "General Resource",
                "description": "General resource representation",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Operation successful"
                    },
                    "data": {
                        "description": "The data returned by the operation",
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "ApiToken": {
                "type": "apiKey",
                "description": "ApiToken required for all end points",
                "name": "Token",
                "in": "header"
            },
            "SanctumBearerToken": {
                "type": "http",
                "description": "Sanctum Bearer Token required after user logs in (Authorization: Bearer <token>)",
                "scheme": "bearer"
            }
        }
    },
...
}